+++ app/src/main/java/com/hacybeyker/finflow/feature/transactions/domain/usecase/GetTransactionsByMonthUseCase.kt
@@ -6,8 +6,7 @@ import java.time.YearMonth
 import javax.inject.Inject
 import kotlinx.coroutines.flow.Flow
 
-/** Reactive list of transactions for a given month. Ordering is left to the data layer. */
-class GetTransactionsByMonth @Inject constructor(private val repository: TransactionRepository) {
+class GetTransactionsByMonthUseCase @Inject constructor(private val repository: TransactionRepository) {
 
     operator fun invoke(month: YearMonth): Flow<List<Transaction>> = repository.observeByMonth(month)
 }